inserido view de compartilhamento de inserido loading no app client

+++ src/AppFidelidade.Dominio/Cliente/Interface/Repositorio/ICompraRepositorio.cs
@@ -6,7 +6,7 @@ namespace AppFidelidade.Dominio.Cliente.Interface.Repositorio
 {
     public interface ICompraRepositorio : IBaseRepositorio<Entidade.Compra>
     {
-        Entidade.Compra ObterPorId(int id, string[] includes);
+        Entidade.Compra ObterPorId(long id, string[] includes);
         List<Entidade.Compra> ObterPorFilial(int idFilial, string[] includes);
         List<Entidade.Compra> ObterPorCliente(int idCliente, string[] includes);
         List<Entidade.Compra> ObterComCreditoPorCliente(int idCliente, string[] includes);
